Move nested mock data utilities out of `test-utils`
